refactor: create a separate file for btc_client
